requirements.txt
requests
pytest
psycopg2
allure-pytest
pytest-html
